+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/internal/Core/OpenCV/VisionProcessors/BrownGlyphFinder.kt
@@ -1,15 +1,17 @@
 @file:Suppress("PackageDirectoryMismatch")
 package org.directcurrent.opencv.visionprocessors
 
+
+import org.directcurrent.opencv.equalizeIntensity
 import org.opencv.core.*
 import org.opencv.imgproc.Imgproc
+import java.lang.Math.abs
 
 
 class BrownGlyphFinder: VisionProcessor()
 {
-    private var limitedHsvMat: Mat? = null
-    private var contourMat: Mat? = null
-    private var boundingMat: Mat? = null
+    private var displayMat: Mat? = null
+    private var processingMat: Mat? = null
 
 
     /**
@@ -17,9 +19,8 @@ class BrownGlyphFinder: VisionProcessor()
      */
     override fun initMats(width: Int, height: Int)
     {
-        limitedHsvMat = Mat(width , height , CvType.CV_8UC4)
-        contourMat = Mat(width , height , CvType.CV_8UC4)
-        boundingMat = Mat(width , height , CvType.CV_8UC4)
+        displayMat = Mat(width , height , CvType.CV_8UC4)
+        processingMat = Mat(width , height , CvType.CV_8UC4)
     }
 
 
@@ -28,56 +29,82 @@ class BrownGlyphFinder: VisionProcessor()
      */
     override fun processFrame(originalMat: Mat?): Mat?
     {
-        limitedHsvMat = Mat()
-        originalMat?.copyTo(contourMat)
-        originalMat?.copyTo(boundingMat)
+        originalMat?.copyTo(displayMat)
 
+
+        // Equalize using Histogram Equalization
+        processingMat = equalizeIntensity(originalMat)
+        originalMat?.release()
+
+        // Blur for extra edge detection
+        Imgproc.medianBlur(processingMat , processingMat , 5)
 
         /*
          * Get our HSV mat, and limit restrict it to a color range we want
          * Color range is HSV- you can get this by taking a picture with the phone and running
          * it through GRIP
          */
-        Imgproc.cvtColor(originalMat , limitedHsvMat , Imgproc.COLOR_RGB2HSV)
-        Core.inRange(limitedHsvMat , Scalar(0.0 , 37.0 , 69.0) ,
-                Scalar(16.0 , 179.0 , 142.0) , limitedHsvMat)
+        Imgproc.cvtColor(processingMat , processingMat , Imgproc.COLOR_RGB2HSV)
+        Core.inRange(processingMat , Scalar(0.0 , 85.0 , 32.0) ,
+                Scalar(25.0 , 196.0 , 140.0) , processingMat)
 
 
-        Imgproc.erode(limitedHsvMat , limitedHsvMat , Imgproc.getStructuringElement
+        Imgproc.erode(processingMat , processingMat , Imgproc.getStructuringElement
+        (Imgproc.MORPH_RECT, Size(2.0 , 2.0)) , Point(0.0 , 0.0) , 1)
+
+        Imgproc.dilate(processingMat , processingMat , Imgproc.getStructuringElement
         (Imgproc.MORPH_RECT, Size(2.0 , 2.0)) , Point(0.0 , 0.0) , 3)
-
-        Imgproc.dilate(limitedHsvMat , limitedHsvMat , Imgproc.getStructuringElement
-        (Imgproc.MORPH_RECT, Size(2.0 , 2.0)) , Point(0.0 , 0.0) , 5)
 
 
         // Declare a list of contours, find them, and then draw them.
         val contours = ArrayList<MatOfPoint>()
         val filteredContours = ArrayList<MatOfPoint>()
 
-        Imgproc.findContours(limitedHsvMat , contours , Mat() , Imgproc.RETR_EXTERNAL , Imgproc.CHAIN_APPROX_SIMPLE)
+        Imgproc.findContours(processingMat , contours , Mat() , Imgproc.RETR_EXTERNAL , Imgproc.CHAIN_APPROX_SIMPLE)
+        processingMat?.release()
 
         // Filter out some contours
-        contours.filterTo(filteredContours) { Imgproc.contourArea(it) >= 2_000 }
+        contours.filterTo(filteredContours)
+        {
+            Imgproc.contourArea(it) >= 22_500
+        }
 
-        Imgproc.drawContours(contourMat , filteredContours , -1 , Scalar(255.0 , 0.0 , 0.0) , 6)
+        displayInfo(filteredContours)
 
-        // Draw bounding rectangles over contours
-        filteredContours
-                .map { Imgproc.boundingRect(it) }
-                .forEach {
-                    Imgproc.rectangle(boundingMat , Point(it.x.toDouble() , it.y.toDouble()) ,
-                            Point((it.x + it.width).toDouble(), (it.y + it.height).toDouble()) ,
+        return displayMat
+    }
+
+
+    override fun displayInfo(contours: ArrayList<MatOfPoint>)
+    {
+        for(i in contours)
+        {
+            val rect = Imgproc.boundingRect(i)
+
+            // If it's roughly in the shape of a square
+            if(abs(rect.height - rect.width) <= 50 && rect.height >= 125 && rect.width >= 125)
+            {
+                if(rect.x >= 500)
+                {
+                    Imgproc.rectangle(displayMat, Point(rect.x.toDouble() , rect.y.toDouble()) ,
+                            Point((rect.x + rect.width).toDouble() , (rect.y + rect.height).toDouble()) ,
                             Scalar(255.0 , 0.0 , 0.0) , 6)
+
+                    Imgproc.putText(displayMat, "Brown Glyph :)"
+                            , Point(rect.x.toDouble() ,
+                            (rect.y + rect.height + 50).toDouble()) , Core.FONT_HERSHEY_COMPLEX , 1.0 ,
+                            Scalar(0.0 , 255.0 , 0.0) , 3)
+
+                    Imgproc.putText(displayMat, "x: " + rect.x , Point(rect.x.toDouble() ,
+                            (rect.y + rect.height + 90).toDouble()) , Core.FONT_HERSHEY_COMPLEX , 1.0 ,
+                            Scalar(0.0 , 255.0 , 0.0) , 3)
+
+                    Imgproc.putText(displayMat, "y: " + rect.y , Point(rect.x.toDouble() ,
+                            (rect.y + rect.height + 130).toDouble()) , Core.FONT_HERSHEY_COMPLEX , 1.0 ,
+                            Scalar(0.0 , 255.0 , 0.0) , 3)
                 }
-
-
-
-        // Deleting pointers
-        originalMat?.release()
-        limitedHsvMat?.release()
-        contourMat?.release()
-
-        return boundingMat
+            }
+        }
     }
 
 
@@ -86,8 +113,6 @@ class BrownGlyphFinder: VisionProcessor()
      */
     override fun releaseMats()
     {
-        limitedHsvMat?.release()
-        contourMat?.release()
-        boundingMat?.release()
+        displayMat?.release()
     }
 }